amend tests and sql

+++ loginservice/src/main/resources/database/login_data.sql
@@ -1,5 +1,6 @@
 USE login;
 
-DELETE FROM User;
+DELETE FROM User WHERE id = 1;
+DELETE FROM User WHERE email = 'admin@admin.com';
 
 INSERT INTO `User` VALUES (1, 'admin@admin.com', 'admin', 'hr');
