+++ processservice/src/main/resources/database/process_data.sql
@@ -1,5 +1,6 @@
 USE process;
 
-DELETE FROM User;
+DELETE FROM User WHERE id = '1';
+DELETE FROM User WHERE name = 'admin@admin.com';
 
 INSERT INTO `User` VALUES ('1', 'admin@admin.com', 'hr');
